Added testing suite

+++ server/src/Bananagrams.Service/Services/UserService.cs
@@ -3,7 +3,9 @@ using Bananagrams.Dal.Interfaces;
 using Bananagrams.Dal.Models;
 using Bananagrams.Dal.Specifications.Users;
 using Bananagrams.Service.Dtos.Users;
+using Bananagrams.Service.Exceptions;
 using Bananagrams.Service.Interfaces;
+using Microsoft.EntityFrameworkCore;
 using Unosquare.EntityFramework.Specification.Common.Extensions;
 
 namespace Bananagrams.Service.Services;
@@ -16,46 +18,57 @@ public class UserService : IUserService
     public UserService(IBananagramsDatabase database, IMapper mapper) =>
         (_database, _mapper) = (database, mapper);
 
-    public IList<UserDto> GetAll(string? searchWord = null)
+    public async Task<IList<UserDto>> GetAll(string? searchWord = null)
     {
-        var users = _mapper.ProjectTo<UserDto>(_database.Get<User>()
+        var users = await _mapper.ProjectTo<UserDto>(_database.Get<User>()
             .Where( new UserByFirstNameSpec(searchWord).Or( new UserByLastNameSpec(searchWord) ).Or( new UserByUsernameSpec(searchWord) ) )
             )
-            .ToList();
+            .ToListAsync();
         
         return users;
     }
 
-    public UserDto Get(int id)
+    public async Task<UserDto> Get(int id)
     {
-        var viewModel = _mapper.ProjectTo<UserDto>(_database.Get<User>()
+        var user = await _mapper.ProjectTo<UserDto>(_database.Get<User>()
             .Where(new UserByIdSpec(id)))
-            .SingleOrDefault();
+            .SingleOrDefaultAsync();
 
-        return viewModel;
+        return user ?? throw new NotFoundException($"Could not find user with id: {id}");
     }
 
-    public void Create(UserDto user)
+    public async Task Create(CreateUserDto user)
     {
         var newUser = _mapper.Map<User>(user);
 
+        newUser.Password = BCrypt.Net.BCrypt.HashPassword(newUser.Password);
+
         _database.Add(newUser);
-        _database.SaveChanges();
+        await _database.SaveChangesAsync();
     }
 
-    public void Update(int id, UserDto user)
+    public async Task Update(int id, UpdateUserDto user)
     {
         var existingUser = _database.Get<User>().FirstOrDefault( new UserByIdSpec(id));
+        
+        if (existingUser == null) throw new NotFoundException($"Could not find user with id: {id}");
+
+        if (user.Password != null)
+        {
+            user.Password = BCrypt.Net.BCrypt.HashPassword(user.Password);
+        }
+        
         _mapper.Map(user, existingUser);
         
-        _database.SaveChanges();
+        await _database.SaveChangesAsync();
     }
 
-    public void Delete(int id)
+    public async Task Delete(int id)
     {
         var existingUser = _database.Get<User>().FirstOrDefault( new UserByIdSpec(id));
-        
+        if (existingUser == null) throw new NotFoundException($"Could not find user with id: {id}");
+
         _database.Delete(existingUser);
-        _database.SaveChanges();
+        await _database.SaveChangesAsync();
     }
 }
